+++ src/binding/petsc4py/src/petsc4py/PETSc/Log.pyx
@@ -168,29 +168,6 @@ cdef class Log:
         CHKERR(PetscLogFlops(cflops))
 
     @classmethod
-    def addFlops(cls, flops: float) -> None:
-        """Add floating point operations to the current event.
-
-        Not collective.
-
-        Parameters
-        ----------
-        flops
-            The number of flops to log.
-
-        Notes
-        -----
-        This method exists for backward compatibility.
-
-        See Also
-        --------
-        logFlops, petsc.PetscLogFlops
-
-        """
-        cdef PetscLogDouble cflops=flops
-        CHKERR(PetscLogFlops(cflops))
-
-    @classmethod
     def getFlops(cls) -> float:
         """Return the number of flops used on this processor since the program began.