+++ src/binding/petsc4py/test/test_dmda.py
@@ -286,8 +286,8 @@ class BaseTestDA:
 
     def testzeroRowsColumnsStencil(self):
         da = self.da
-        A = da.createMatrix()
-        x = da.createGlobalVector()
+        A = da.createMat()
+        x = da.createGlobalVec()
         x.set(2.0)
         A.setDiagonal(x)
         diag1 = x.duplicate()